# LocalCudaWorkerService.Client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "LocalCudaWorkerService API",
    "description": "ASP.NET REST-API using CUDA as an online worker for executing kernels via HTTP.",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "github: alarmclock-kisser",
      "email": "marcel.king91299@gmail.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Test/test-cufft-forward-single": {
      "get": {
        "tags": [
          "Test"
        ],
        "parameters": [
          {
            "name": "fftSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "deviceName",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "NVIDIA"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/float2"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/float2"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/float2"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Test/test-cufft-inverse-single": {
      "get": {
        "tags": [
          "Test"
        ],
        "parameters": [
          {
            "name": "fftSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "deviceName",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "NVIDIA"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "number",
                    "format": "float"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "number",
                    "format": "float"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "number",
                    "format": "float"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Test/test-execute-single-round": {
      "get": {
        "tags": [
          "Test"
        ],
        "parameters": [
          {
            "name": "deviceName",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "NVIDIA"
            }
          },
          {
            "name": "inputSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1024
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/KernelExecuteResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KernelExecuteResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/KernelExecuteResult"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Test/test-execute-batch-round": {
      "get": {
        "tags": [
          "Test"
        ],
        "parameters": [
          {
            "name": "deviceName",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "NVIDIA"
            }
          },
          {
            "name": "inputSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1024
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/KernelExecuteResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KernelExecuteResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/KernelExecuteResult"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "KernelExecuteResult": {
        "type": "object",
        "properties": {
          "kernelName": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "outputPointer": {
            "type": "string",
            "nullable": true
          },
          "outputDataBase64": {
            "type": "string",
            "nullable": true
          },
          "outputDataBase64Chunks": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "outputDataType": {
            "type": "string",
            "nullable": true
          },
          "outputDataLength": {
            "type": "string",
            "nullable": true
          },
          "outputDataStride": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "readOnly": true
          },
          "executionTimeMs": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "float2": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}